General > App Name (#74)
* Added app name configuration

* Added stylings and fixed old stylings

* Added app name to the store

* Increased border radius of VCard

* Removed value prop and change emit from VTextField

* Added scoped style to Configure Sidebar

* Removed wrapper div from AppConfigure

* Moved main tag to GeneralSettings

* Changed border transparent to border none in VSeparator

* Removed unused store elements and added type to state

* Removed wrapper description div

* Updated setting card

* Added wrapper div for description

* Added nested router and routerview to Configure page

* Added flex-grow to all the main elements
