+++ src/components/app-configure/general/GeneralSettings.vue
@@ -1,0 +1,12 @@
+<script lang="ts" setup>
+import AppName from '@/components/app-configure/general/AppName.vue'
+import VStack from '@/components/lib/VStack/VStack.vue'
+</script>
+
+<template>
+  <main>
+    <VStack direction="column">
+      <AppName />
+    </VStack>
+  </main>
+</template>
