+++ src/components/lib/VSeperator/VSeperator.vue
@@ -15,10 +15,18 @@ defineProps({
 <style scoped>
 hr,
 .vr {
-  filter: blur(1px);
-  border: 1px solid #000;
+  border: none;
   border-radius: 10px;
   box-shadow: 1px 1px 3px 4px rgb(45 45 45 / 8%),
     inset -10px -10px 20px rgb(0 0 0 / 50%), inset 10px 10px 15px #222;
+}
+
+hr {
+  width: 100%;
+  border-top: 1px solid #000;
+}
+
+.vr {
+  border-left: 1px solid #000;
 }
 </style>
